network:bluetooth: Share address functions
Some functions to deal with bluetooth address can be shared between all
platforms.
